Debian: Add post install script
This DKMS script will be run when the package is installed.
It will run the standard DKMS build steps (compiling the module against
the current kernel version), and will additionally:

- Load the module immediately
- Configure the module to be loaded automatically at boot

This is not handled by default by DKMS, because usually, kernel modules
are associated with some hardware, and the hardware detection in the
kernel will be enough to load the module.

But, in the case of tty0tty, there is no associated hardware, and so,
the module must be explicitly loaded.